systemd-journal: make the journal source is the system default
 * if /dev/log source is set in the config and systemd is present,
   syslog-ng fallback to systemd-journal source
 * if the service manager is systemd based, then load the sdjournal plugin automatically

+++ modules/systemd-journal/journal-reader.c
@@ -159,7 +159,7 @@ __handle_data(gchar *key, gchar *value, gpointer user_data)
     {
       log_msg_set_value(msg, LM_V_MESSAGE, value, value_len);
       msg_debug("Incoming log entry from journal",
-                evt_tag_printf("message", "%.*s", value_len, value),
+                evt_tag_printf("message", "%.*s", (int)value_len, value),
                 NULL);
     }
   else if (strcmp(key, "_HOSTNAME") == 0)
